+++ source/testing/CXXBrusselatorExecutableModel.h
@@ -67,7 +67,7 @@ public:
     virtual int getNumFloatingSpecies();
 
     virtual int getFloatingSpeciesIndex(const std::string& eid);
-    virtual std::string getFloatingSpeciesId(int index);
+    virtual std::string getFloatingSpeciesId(size_t index);
 
     /**
      * independent species do are not defined by rules, they typically participate
@@ -174,7 +174,7 @@ public:
      */
     virtual int getNumBoundarySpecies();
     virtual int getBoundarySpeciesIndex(const std::string &eid);
-    virtual std::string getBoundarySpeciesId(int index);
+    virtual std::string getBoundarySpeciesId(size_t index);
 
     /**
      * get the boundary species amounts
@@ -233,7 +233,7 @@ public:
     /**
      * id of the indexed global parameter.
      */
-    virtual std::string getGlobalParameterId(int index);
+    virtual std::string getGlobalParameterId(size_t index);
 
     /**
      * get the global parameter values
@@ -261,7 +261,7 @@ public:
 
     virtual int getNumCompartments();
     virtual int getCompartmentIndex(const std::string& eid);
-    virtual std::string getCompartmentId(int index);
+    virtual std::string getCompartmentId(size_t index);
 
     /**
      * get the compartment volumes
@@ -360,7 +360,7 @@ public:
 
     virtual int getNumConservedMoieties();
     virtual int getConservedMoietyIndex(const std::string& eid);
-    virtual std::string getConservedMoietyId(int index);
+    virtual std::string getConservedMoietyId(size_t index);
     virtual int getConservedMoietyValues(size_t len, int const *indx, double *values);
     virtual int setConservedMoietyValues(size_t len, int const *indx,
             const double *values);
@@ -381,7 +381,7 @@ public:
     /**
      * get the name of the specified reaction
      */
-    virtual std::string getReactionId(int index);
+    virtual std::string getReactionId(size_t index);
 
     /**
      * get the vector of reaction rates.
@@ -406,7 +406,7 @@ public:
     /**
      * get the id of an element of the state vector.
      */
-    virtual std::string getStateVectorId(int index);
+    virtual std::string getStateVectorId(size_t index);
 
     /**
      * The state vector is a vector of elements that are defined by
@@ -515,9 +515,9 @@ public:
      * If there is no event with this id, returns -1.
      */
     virtual int getEventIndex(const std::string& eid);
-    virtual std::string getEventId(int index);
-    virtual void setEventListener(int index, rr::EventListenerPtr eventHandler);
-    virtual rr::EventListenerPtr getEventListener(int index);
+    virtual std::string getEventId(size_t index);
+    virtual void setEventListener(size_t index, rr::EventListenerPtr eventHandler);
+    virtual rr::EventListenerPtr getEventListener(size_t index);
 
     /**
      * Get the amount rate of change for the i'th floating species
@@ -532,7 +532,7 @@ public:
      * @param index: index of the desired floating speceis rate.
      * @param reactionRates: pointer to buffer of reaction rates.
      */
-    virtual double getFloatingSpeciesAmountRate(int index,
+    virtual double getFloatingSpeciesAmountRate(size_t index,
             const double *reactionRates);
 
     /**
